Enhance Azure OpenAI integration and refactor audio utilities
- Added new Azure OpenAI configuration variables in config.py for LightRAG LLM integration.
- Updated main.py to utilize the new llm_model_func for improved model handling.
- Refactored audio utilities by moving get_json_dict_from_audio to litellm_utils.py for better organization.
- Improved the ContextualChunkETLPipeline to handle transcript updates more effectively.
- Renamed test function for consistency in unit tests.
