+++ echo/server/dembrane/config.py
@@ -140,10 +140,26 @@ logger.debug("NEO4J_PASSWORD: set")
 
 
 # Azure OpenAI configurations
+
+AZURE_OPENAI_LIGHTRAGLLM_API_KEY = os.environ.get("AZURE_OPENAI_LIGHTRAGLLM_API_KEY")   
+assert AZURE_OPENAI_LIGHTRAGLLM_API_KEY, "AZURE_OPENAI_LIGHTRAGLLM_API_KEY environment variable is not set"
+logger.debug("AZURE_OPENAI_LIGHTRAGLLM_API_KEY: set")
+
+AZURE_OPENAI_LIGHTRAGLLM_API_VERSION = os.environ.get("AZURE_OPENAI_LIGHTRAGLLM_API_VERSION")
+assert AZURE_OPENAI_LIGHTRAGLLM_API_VERSION, "AZURE_OPENAI_LIGHTRAGLLM_API_VERSION environment variable is not set"
+logger.debug("AZURE_OPENAI_LIGHTRAGLLM_API_VERSION: set")
+
+AZURE_OPENAI_LIGHTRAGLLM_ENDPOINT = os.environ.get("AZURE_OPENAI_LIGHTRAGLLM_ENDPOINT")
+assert AZURE_OPENAI_LIGHTRAGLLM_ENDPOINT, "AZURE_OPENAI_LIGHTRAGLLM_ENDPOINT environment variable is not set"
+logger.debug("AZURE_OPENAI_LIGHTRAGLLM_ENDPOINT: set")
+
+AZURE_OPENAI_LIGHTRAGLLM_NAME = os.environ.get("AZURE_OPENAI_LIGHTRAGLLM_NAME")
+assert AZURE_OPENAI_LIGHTRAGLLM_NAME, "AZURE_OPENAI_LIGHTRAGLLM_NAME environment variable is not set"
+logger.debug("AZURE_OPENAI_LIGHTRAGLLM_NAME: set")
+
 AZURE_OPENAI_API_KEY = os.environ.get("AZURE_OPENAI_API_KEY")
 assert AZURE_OPENAI_API_KEY, "AZURE_OPENAI_API_KEY environment variable is not set"
 logger.debug("AZURE_OPENAI_API_KEY: set")
-
 AZURE_OPENAI_API_VERSION = os.environ.get("AZURE_OPENAI_API_VERSION")
 assert AZURE_OPENAI_API_VERSION, "AZURE_OPENAI_API_VERSION environment variable is not set"
 logger.debug("AZURE_OPENAI_API_VERSION: set")
